Modify detail pages to use AddButtonGroup
This makes the detail pages more consistent, and allows changes to all
of the Add button group to happen in one location. Some of the files
previously used three columns to simulate an add button group rather
than using the actual view, so I was able to simplify much of the xaml.