Move our popups to use popup v2 MCT
This does introduce at least one bug, where adding a new book in the
book searcher utility will reopen the page, including the underlying
BookDetailPage. This is probably a shell issue, or the popup itself is
reopeing the page, even though it should have been open in the first
place.